refactor: boards/index, BestArticles
