+++ components/Layout/Header.tsx
@@ -1,4 +1,4 @@
-import Logo from "../assets/images/logo/logo.svg";
+import Logo from "../../public/images/logo/logo.svg";
 import React from "react";
 import { useRouter } from "next/router";
 import Image from "next/image";
